TVGUI-1 Converting RecordedShow page to TS
